Added user validation and authorization to Routine and Goal subsystems; refactored Routine delete stack to run based off a provided int ID instead of an entire routine object

+++ server/src/main/java/learn/fitness/domain/GoalService.java
@@ -53,7 +53,11 @@ public class GoalService {
         return result;
     }
 
-    public boolean delete(int goalId) {
+    public boolean delete(int goalId, int userId) {
+        if(findByUser(userId).stream().noneMatch(g -> g.getGoal_id() == goalId)){
+            return false; // Goal does not exist or does not belong to user; treat as not found
+        }
+
         return repository.delete(goalId);
     }
